Document namespace merge ordering and relocate guard test to Integration
